Updated the SQL script. Modified the Notifications table since it was missing primary key and a foreign key reference.
Updated the DBNotification class and Notification Interface to use long
for ItemID.

Implemented DBNotificationQuery which add, gets, removes, and updates
notifications. Also implemented the test class to run a JUnit test.

+++ src/main/java/edu/cpp/cs580/Database/Objects/Interfaces/Notification.java
@@ -4,11 +4,11 @@ import java.math.BigInteger;
 
 public interface Notification {
 	public double getCurrentPrice();
-	public BigInteger getItemID();
+	public long getItemID();
 	public double getNotifyPrice();
 	public String getUsername();
 	public void setCurrentPrice(double price);
-	public void setItemID(BigInteger id);
+	public void setItemID(long id);
 	public void setNotifyPrice(double price);
 	public void setUsername(String name);
 }
